ml_service/requirements.txt
# ML Service Requirements
# Python 3.9+

# Web Framework
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0

# ML/AI Libraries
scikit-learn>=1.3.0
joblib>=1.3.0
numpy>=1.24.0
scipy>=1.10.0

# Environment & Configuration
python-dotenv>=1.0.0

# HTTP Client (for backend integration)
requests>=2.28.0

# MQTT Client (for publishing recommendations)
paho-mqtt>=1.6.1

# Database clients (for storing predictions)
influxdb>=5.3.1
psycopg2-binary>=2.9.6

# Logging & Utilities
python-json-logger>=2.0.0

